+++ DaySim/DaySim.csproj
@@ -265,6 +265,7 @@
     <Compile Include="DestinationParkingShadowPricing\DestinationParkingShadowPriceReader.cs" />
     <Compile Include="DestinationParkingShadowPricing\DestinationParkingShadowPriceWriter.cs" />
     <Compile Include="DestinationParkingDao.cs" />
+    <Compile Include="PathTypeModels\PathTypeModel_Actum.cs" />
     <Compile Include="PathTypeModels\PathTypeModel_override.cs" />
     <Compile Include="PathTypeModels\IPathTypeModel.cs" />
     <Compile Include="PathTypeModels\PathTypeModel.cs" />
